requirements.txt
django
djangorestframework
django-cors-headers
Pillow
httpie

